msm: smd_tty: Introduce port-specific wait while opening channel
Currently SMD TTY uses a single parameter for the timeout value that is
shared across all SMD ports. However, each SMD TTY port has unique
constraints on the timeout value as the timeout depends on the subsystem
that the port is connecting to.

Add a device attribute that provides a timeout value for each port.

Change-Id: I0934d6b164b8b9e90917fab1859493478076dc83
Signed-off-by: Zaheerulla Meer <zmeer@codeaurora.org>