arm/dt: msm8974: Remove PLL commands for Standalone PC
On V2 hardware, the SPM can handle turning on/off HFPLL only if the core
was running on HFPLL before stand alone power collapse. This operations
requires a 50us delay for the HFPLL to stabilize when entering
standalone power collapse. This results in higher power numbers when the
device is in a idle state. So remove the PLL on/off commands for
standalone power collapse only.

Change-Id: I517fb9aeade6084d6d0ad07726bd9e6b10d83bfb
Signed-off-by: Mahesh Sivasubramanian <msivasub@codeaurora.org>