msm: clock-8974: Add enable mask for gpll0 clock
gpll0 clock's enable mask was removed to avoid gpll0 getting
turned off while in use in pre-silicon simulation environments.
However, with the bootchain environments on hardware, this
does not happen anymore because:

   1) The bootchain and RPM will vote for GPLL0 to remain enabled when
      the Apps is running.
   2) Clock driver will vote for GPLL0 when there are clocks clients
      requiring it.

Note that acpuclock not voting for GPLL0 is not a problem due to
point (1) above.

Therefore, add the enable mask back for gpll0.

Change-Id: I85d79371ba7d2118c4cfa9ecfb3f453eef10d696
Signed-off-by: Tianyi Gou <tgou@codeaurora.org>