msm: camera: Do not do HW reset of the ISPIF
There is no real need to reset the ISPIF, as all the ISPIF reset is
doing is to reset the registers to default values. But there are too
many problems trying to do a reset of the ISPIF as it requires all the
clocks from CSID and all clocks from every VFE interface to complete the
reset. Because of this it simpler to replace the HW reset with a SW one,
where the registers are set to the default value.

The only place where ISPIF driver explicitly requires clocks to be
enabled is when reset is performed. Because the ISPIF needs only the
ispif_ahb_clk clock to be programmed there is also no need to keep the
request for the other clocks in the driver. So we can leave this job to
the CSID and VFE blocks themselves.

This way only the ispif_ahb_clk clock enable requests will be done by
the ISPIF and the other clock enable requests will be left to the CSID
and VFE drivers per device only when they are used and really needed.
This will be enough for the ISPIF to work as desired, as there is no
time when the ISPIF can work alone without the CSID and/or VFE running
in the same time.

Change-Id: I0dbde8b4fef6ee4d5ff15f2195b3dd7936d5eb07
Signed-off-by: Vladislav Hristov <vhrist@codeaurora.org>