mtd: msm_qpic_nand: Use address from APPS partition for ONFI probe
The ONFI device probe involves requesting the device for ONFI params.
This request is sent as a page-read with zeroth address (from MTD
partition table) as the address used is considered DONT CARE by the
device. But MPU/XPU would flag it Nand Access violation as that page
address may not be accessible by APPS proc

So instead pass the page address of 'APPS' partition on which APPS proc
has read/write permissions.

CRs-Fixed: 465463
Change-Id: I9f1715c0385fa7732e2879417305ebfdecbf8679
Signed-off-by: Pratibhasagar V <pratibha@codeaurora.org>