regulator: qpnp-regulator: Add support for voltage switch OCP IRQ
Add support for a voltage switch OCP interrupt handler which
toggles the switch off and back on after an overcurrent event
takes place.  This toggling should only be performed for a
specified maximum number of times with a specified delay between
subsequent retries.  Disabling a voltage switch is the only way
to clear the latched over-current state.

Change-Id: I8dbc54ff7bc54c1309ee59c1a282d5e4af2f0ec6
CRs-Fixed: 469134
Signed-off-by: David Collins <collinsd@codeaurora.org>