usb: dwc3-msm: Notify MHL of cable state at bootup
Commit f08f7b67 "usb: dwc3-msm: Read ID line state on bootup"
fixed a bug in which a USB micro-A cable plugged in during bootup
would fail to be detected. However MHL cables still fail to be
detected at bootup since the MHL driver never gets a chance to
perform its cable discovery and thus gets misidentified as a USB
host cable. Fix this by processing the initial ID low state in the
ID work function which calls the MHL discovery mechanism first.

CRs-fixed: 469986
Change-Id: I9c54a590c3e5f3574fa51d0a190d047135605e60
Signed-off-by: Jack Pham <jackp@codeaurora.org>