media: dvb: Move video pattern search to dvb-demux level
Video pattern search is required for both video framing
done at plug-in level and indexing to be done by
dvb-demux level. The pattern search routine was moved as
a common service at dvb-demux level so that it can be
used by both.

At plug-in level, the plug-in needs the video codec in
order to perform the framing regardless of the indexing
parameters, therefore indexing parameters was renamed
to video codec.

Change-Id: I2e3e648c3a7d09254e7284c25ca7550089887907
Signed-off-by: Hamad Kadmany <hkadmany@codeaurora.org>