msm: smem: Add virtual-to-physical mapping
Some users of SMEM items need to pass the item address to a DMA engine.
This requires a physical address.  Since SMEM is mapped as IO memory,
the normal virt_to_phys() and dma_map functions cannot be used.

Add function to SMEM that does the virtual to physical mapping.

CRs-Fixed: 474248
Change-Id: I528b58dea3c81d9ec3f270b76275f883f7145887
Signed-off-by: Eric Holmberg <eholmber@codeaurora.org>