msm: camera: Return buffer to queue when hardware stops
When ISP hardware is active, it acquire two image buffers to
configure the hardware ping pong buffer. When hardware is stopped,
the buffer are not put back to the queue. This change add the support
to put the buffer back to the queue in order for image stream to restart
without recreating a new buffer queue.

Change-Id: I2da5db0d4c258be9f19133312b39d086ceeeeaee
Signed-off-by: Kevin Chan <ktchan@codeaurora.org>