mtd: msm_qpic_nand: Check for the updated partition name
The partition name when read from SMEM is available in raw
format. They are not plain text but are prefixed by other
data and delimiter. If we do a lookup for partition name in it
the check may fail and may lead to wrong page address.

The partition name is updated and saved in another struct. So
instead read the part name from mtd_part(reformatted data) rather
than from the part_entry(raw data).

Change-Id: I9c9b50cdcf600314ab76da21b0f64e02824611b1
Signed-off-by: Pratibhasagar V <pratibha@codeaurora.org>